requirements.txt
streamlit
markdown
xhtml2pdf
python-dotenv
google-generativeai